Small histories (#103) bench: 6225658
lower memory usage!!!!

+++ src/search.h
@@ -68,11 +68,11 @@ struct Engine {
 
         TranspositionTable* TT;
 
-        std::array<std::array<std::array<std::array<std::array<int, 2>, 64>, 2>, 64>, 2> historyTable;
-        std::array<std::array<std::array<std::array<int, 7>, 64>, 7>, 2> noisyHistoryTable;
-        std::array<std::array<std::array<std::array<int, 7>, 64>, 7>, 2> qsHistoryTable;
-        std::array<std::array<std::array<std::array<int, 64>, 7>, 2>, 32768> pawnHistoryTable;
-        std::array<std::array<std::array<int, 2>, 2>, 32768> correctionHistoryTable;
+        std::array<std::array<std::array<std::array<std::array<int16_t, 2>, 64>, 2>, 64>, 2> historyTable;
+        std::array<std::array<std::array<std::array<int16_t, 7>, 64>, 7>, 2> noisyHistoryTable;
+        std::array<std::array<std::array<std::array<int16_t, 7>, 64>, 7>, 2> qsHistoryTable;
+        std::array<std::array<std::array<std::array<int16_t, 64>, 7>, 2>, 32768> pawnHistoryTable;
+        std::array<std::array<std::array<int32_t, 2>, 2>, 32768> correctionHistoryTable;
         std::unique_ptr<CHTable> conthistTable;
         std::array<std::array<Move, 64>, 64> counterMoves;
